refactor(agent): extract prompt input select wrappers
